player animation
player is now properly animated, which also fixed dashes going in only 1 direction. State machine is also now set up to player animations if available.

How To Use: State machine know has an export variable that receive an AnimationPlayer node that will automatically player the corresponding animation once it enters a state. the animation simply needs to be the same name as the corresponding state and in all caps.